MSH5-20-station-filters-dropdown-menus-component: Moved filter components to a separate filters folder
